refactor: rename `Config` to `ResolvedConfig`
